fix recording only record 1sec
When recording AutoCAD UI auto testing, we found the record stops after
1 secs, after dumping errors in delegates. I found sometimes it stops
with -11809 error (source media changed).
I found specifying the Preset to a fixed size can fix this problem.
(ideally, I should be able to set the source format to be same as
screen size, but 1920x1080 is not available on Mac)